+++ src/components/pitel_sdk.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Platform, LogBox } from 'react-native';
 import { getFcmToken, NotificationListener } from '../notification/push_notif';
-import { encodeDisplayName } from '../utils/display_name_helper';
 import {
   registerDeviceToken,
   removeDeviceToken,
@@ -48,11 +47,6 @@ export const PitelSDK = ({
       iceServers.push(turn.data);
     }
 
-    const displayName = encodeDisplayName({
-      displayNameRaw: sdkOptionsInit.displayName,
-      phoneNumber: sdkOptionsInit.extension,
-    });
-
     const sdkOptionsInitialize = {
       sipOnly: true,
       sipDomain: `${sdkOptionsInit.sipDomain}:${sdkOptionsInit.port}`,
@@ -62,7 +56,10 @@ export const PitelSDK = ({
       debug: true,
       contactName: sdkOptionsInit.extension,
       viaHost: sdkOptionsInit.sipDomain,
-      displayName: displayName,
+      displayName:
+        sdkOptionsInit.displayName != ''
+          ? sdkOptionsInit.displayName
+          : sdkOptionsInit.extension,
       preloadedRouteSet: [
         `Route: <sip:${sdkOptionsInit.sipDomain}:${sdkOptionsInit.port};lr;sipml5-outbound;transport=udp>`,
       ],